@@ -25,17 +25,17 @@
     "@otplib/plugin-base32-enc-dec": "^12.0.1",
     "@otplib/plugin-crypto-js": "^12.0.1",
     "@otplib/preset-browser": "^12.0.1",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@sveltestrap/sveltestrap": "^6.2.7",
     "@tsconfig/svelte": "^5.0.0",
     "@types/qrcode": "^1.5.0",
     "@types/ua-parser-js": "^0.7.36",
-    "@typescript-eslint/eslint-plugin": "^7.17.0",
-    "@typescript-eslint/parser": "^7.18.0",
+    "@xterm/addon-serialize": "^0.13",
+    "@xterm/xterm": "^5.5",
     "bootstrap": "^5.3.3",
     "copy-text-to-clipboard": "^3.0.1",
     "date-fns": "^4.1.0",
-    "eslint": "^8",
+    "eslint": "^9.13.0",
     "eslint-config-standard": "^17.1.0",
     "eslint-import-resolver-typescript": "^3.6.3",
     "eslint-plugin-import": "^2.31.0",
@@ -43,9 +43,10 @@
     "eslint-plugin-promise": "^7.1.0",
     "eslint-plugin-svelte": "^2.46.0",
     "format-duration": "^3.0.2",
+    "otpauth": "^9.3.4",
     "qrcode": "^1.5.4",
-    "sass": "^1.80.4",
-    "svelte": "^4.2.19",
+    "sass": "~1.78",
+    "svelte": "^5.1.3",
     "svelte-check": "^4.0.5",
     "svelte-fa": "^4.0.3",
     "svelte-intersection-observer": "^1.0.0",
@@ -58,9 +59,10 @@
     "ua-parser-js": "^1.0.39",
     "vite": "^5.4.10",
     "vite-plugin-checker": "^0.8.0",
-    "vite-tsconfig-paths": "^4.3.2",
-    "@xterm/xterm": "^5.5",
-    "@xterm/addon-serialize": "^0.13",
-    "otpauth": "^9.3.4"
+    "vite-tsconfig-paths": "^4.3.2"
+  },
+  "dependencies": {
+    "@stylistic/eslint-plugin": "^2.9.0",
+    "typescript-eslint": "^8.11.0"
   }
 }
