@@ -1,8 +1,12 @@
 <script lang="ts">
-import { type Tls, TlsMode } from 'admin/lib/api'
-import { FormGroup, Input } from '@sveltestrap/sveltestrap'
+    import { type Tls, TlsMode } from 'admin/lib/api'
+    import { FormGroup, Input } from '@sveltestrap/sveltestrap'
 
-export let value: Tls
+    interface Props {
+        value: Tls;
+    }
+
+    let { value = $bindable() }: Props = $props()
 
 </script>
 
