@@ -1,247 +1,257 @@
 <script lang="ts">
-import { faIdBadge, faKey, faKeyboard, faMobileScreen } from '@fortawesome/free-solid-svg-icons'
-import { api, CredentialKind, type Role, type User, type UserAuthCredential, type UserRequireCredentialsPolicy } from 'admin/lib/api'
-import AsyncButton from 'common/AsyncButton.svelte'
-import DelayedSpinner from 'common/DelayedSpinner.svelte'
-import Fa from 'svelte-fa'
-import { replace } from 'svelte-spa-router'
-import { Alert, Button, FormGroup, Input } from '@sveltestrap/sveltestrap'
-import AuthPolicyEditor from './AuthPolicyEditor.svelte'
-import UserCredentialModal from './UserCredentialModal.svelte'
-import { stringifyError } from 'common/errors'
+    import { faIdBadge, faKey, faKeyboard, faMobileScreen } from '@fortawesome/free-solid-svg-icons'
+    import { api, CredentialKind, type Role, type User, type UserAuthCredential, type UserRequireCredentialsPolicy } from 'admin/lib/api'
+    import AsyncButton from 'common/AsyncButton.svelte'
+    import DelayedSpinner from 'common/DelayedSpinner.svelte'
+    import Fa from 'svelte-fa'
+    import { replace } from 'svelte-spa-router'
+    import { Button, FormGroup, Input } from '@sveltestrap/sveltestrap'
+    import AuthPolicyEditor from './AuthPolicyEditor.svelte'
+    import UserCredentialModal from './UserCredentialModal.svelte'
+    import { stringifyError } from 'common/errors'
+    import Alert from 'common/Alert.svelte'
 
-export let params: { id: string }
-
-let error: string|null = null
-let user: User
-let editingCredential: UserAuthCredential|undefined
-let policy: UserRequireCredentialsPolicy
-let allRoles: Role[] = []
-let roleIsAllowed: Record<string, any> = {}
-
-const policyProtocols: { id: 'ssh' | 'http' | 'mysql' | 'postgres', name: string }[] = [
-    { id: 'ssh', name: 'SSH' },
-    { id: 'http', name: 'HTTP' },
-    { id: 'mysql', name: 'MySQL' },
-    { id: 'postgres', name: 'PostgreSQL' },
-]
-
-const possibleCredentials: Record<string, Set<CredentialKind>> = {
-    ssh: new Set([CredentialKind.Password, CredentialKind.PublicKey, CredentialKind.Totp, CredentialKind.WebUserApproval]),
-    http: new Set([CredentialKind.Password, CredentialKind.Totp, CredentialKind.Sso]),
-    mysql: new Set([CredentialKind.Password]),
-    postgres: new Set([CredentialKind.Password]),
-}
-
-async function load () {
-    try {
-        user = await api.getUser({ id: params.id })
-        policy = user.credentialPolicy ?? {}
-        user.credentialPolicy = policy
-
-        allRoles = await api.getRoles()
-        const allowedRoles = await api.getUserRoles(user)
-        roleIsAllowed = Object.fromEntries(allowedRoles.map(r => [r.id, true]))
-    } catch (err) {
-        error = await stringifyError(err)
+    interface Props {
+        params: { id: string };
     }
-}
 
-function deleteCredential (credential: UserAuthCredential) {
-    user.credentials = user.credentials.filter(c => c !== credential)
-}
+    let { params }: Props = $props()
 
-function abbreviatePublicKey (key: string) {
-    return key.slice(0, 16) + '...' + key.slice(-8)
-}
+    let error: string|null = $state(null)
+    let user: User | undefined = $state()
+    let editingCredential: UserAuthCredential|undefined = $state()
+    let policy: UserRequireCredentialsPolicy | undefined = $state()
+    let allRoles: Role[] = $state([])
+    let roleIsAllowed: Record<string, any> = $state({})
 
-async function update () {
-    try {
-        user = await api.updateUser({
-            id: params.id,
-            userDataRequest: user,
-        })
-    } catch (err) {
-        error = await stringifyError(err)
+    const policyProtocols: { id: 'ssh' | 'http' | 'mysql' | 'postgres', name: string }[] = [
+        { id: 'ssh', name: 'SSH' },
+        { id: 'http', name: 'HTTP' },
+        { id: 'mysql', name: 'MySQL' },
+        { id: 'postgres', name: 'PostgreSQL' },
+    ]
+
+    const possibleCredentials: Record<string, Set<CredentialKind>> = {
+        ssh: new Set([CredentialKind.Password, CredentialKind.PublicKey, CredentialKind.Totp, CredentialKind.WebUserApproval]),
+        http: new Set([CredentialKind.Password, CredentialKind.Totp, CredentialKind.Sso]),
+        mysql: new Set([CredentialKind.Password]),
+        postgres: new Set([CredentialKind.Password]),
     }
-}
 
-async function remove () {
-    if (confirm(`Delete user ${user.username}?`)) {
-        await api.deleteUser(user)
-        replace('/config')
-    }
-}
+    async function load () {
+        try {
+            user = await api.getUser({ id: params.id })
+            policy = user.credentialPolicy ?? {}
+            user.credentialPolicy = policy
 
-async function toggleRole (role: Role) {
-    if (roleIsAllowed[role.id]) {
-        await api.deleteUserRole({
-            id: user.id,
-            roleId: role.id,
-        })
-        roleIsAllowed = { ...roleIsAllowed, [role.id]: false }
-    } else {
-        await api.addUserRole({
-            id: user.id,
-            roleId: role.id,
-        })
-        roleIsAllowed = { ...roleIsAllowed, [role.id]: true }
+            allRoles = await api.getRoles()
+            const allowedRoles = await api.getUserRoles(user)
+            roleIsAllowed = Object.fromEntries(allowedRoles.map(r => [r.id, true]))
+        } catch (err) {
+            error = await stringifyError(err)
+        }
     }
-}
 
-function saveCredential () {
-    if (!editingCredential) {
-        return
+    function deleteCredential (credential: UserAuthCredential) {
+        user!.credentials = user!.credentials.filter(c => c !== credential)
     }
-    if (user.credentials.includes(editingCredential)) {
-        user.credentials = [...user.credentials]
-    } else {
-        user.credentials.push(editingCredential)
-        for (const protocol of ['http', 'ssh'] as ('http'|'ssh')[]) {
-            for (const ck of [CredentialKind.Password, CredentialKind.PublicKey]) {
-                if (
-                    editingCredential.kind === CredentialKind.Totp
+
+    function abbreviatePublicKey (key: string) {
+        return key.slice(0, 16) + '...' + key.slice(-8)
+    }
+
+    async function update () {
+        try {
+            user = await api.updateUser({
+                id: params.id,
+                userDataRequest: user!,
+            })
+        } catch (err) {
+            error = await stringifyError(err)
+        }
+    }
+
+    async function remove () {
+        if (confirm(`Delete user ${user!.username}?`)) {
+            await api.deleteUser(user!)
+            replace('/config')
+        }
+    }
+
+    async function toggleRole (role: Role) {
+        if (roleIsAllowed[role.id]) {
+            await api.deleteUserRole({
+                id: user!.id,
+                roleId: role.id,
+            })
+            roleIsAllowed = { ...roleIsAllowed, [role.id]: false }
+        } else {
+            await api.addUserRole({
+                id: user!.id,
+                roleId: role.id,
+            })
+            roleIsAllowed = { ...roleIsAllowed, [role.id]: true }
+        }
+    }
+
+    function saveCredential () {
+        if (!editingCredential || !user) {
+            return
+        }
+        if (user.credentials.includes(editingCredential)) {
+            user.credentials = [...user.credentials]
+        } else {
+            user.credentials.push(editingCredential)
+            for (const protocol of ['http', 'ssh'] as ('http'|'ssh')[]) {
+                for (const ck of [CredentialKind.Password, CredentialKind.PublicKey]) {
+                    if (
+                        editingCredential.kind === CredentialKind.Totp
                     && !user.credentialPolicy?.[protocol]
                     && user.credentials.some(x => x.kind === ck)
                     && possibleCredentials[protocol]?.has(ck)
-                ) {
-                    user.credentialPolicy = {
-                        ...user.credentialPolicy ?? {},
-                        [protocol]: [ck, CredentialKind.Totp],
+                    ) {
+                        user.credentialPolicy = {
+                            ...user.credentialPolicy ?? {},
+                            [protocol]: [ck, CredentialKind.Totp],
+                        }
+                        policy = user.credentialPolicy
                     }
-                    policy = user.credentialPolicy
                 }
             }
         }
+        editingCredential = undefined
     }
-    editingCredential = undefined
-}
 
-function assertDefined<T>(value: T|undefined): T {
-    if (value === undefined) {
-        throw new Error('Value is undefined')
+    function assertDefined<T>(value: T|undefined): T {
+        if (value === undefined) {
+            throw new Error('Value is undefined')
+        }
+        return value
     }
-    return value
-}
 </script>
 
 {#await load()}
-    <DelayedSpinner />
+<DelayedSpinner />
 {:then}
-    <div class="page-summary-bar">
-        <div>
-            <h1>{user.username}</h1>
-            <div class="text-muted">User</div>
-        </div>
+{#if user}
+<div class="page-summary-bar">
+    <div>
+        <h1>{user.username}</h1>
+        <div class="text-muted">User</div>
     </div>
+</div>
 
-    <FormGroup floating label="Username">
-        <Input bind:value={user.username} />
-    </FormGroup>
+<FormGroup floating label="Username">
+    <Input bind:value={user.username} />
+</FormGroup>
 
-    <div class="d-flex align-items-center mt-4 mb-2">
-        <h4 class="m-0">Credentials</h4>
+<div class="d-flex align-items-center mt-4 mb-2">
+    <h4 class="m-0">Credentials</h4>
+    <span class="ms-auto"></span>
+    <Button size="sm" color="link" on:click={() => editingCredential = {
+        kind: 'Password',
+        hash: '',
+    }}>Add password</Button>
+    <Button size="sm" color="link" on:click={() => editingCredential = {
+        kind: 'PublicKey',
+        key: '',
+    }}>Add public key</Button>
+    <Button size="sm" color="link" on:click={() => editingCredential = {
+        kind: 'Totp',
+        key: [],
+    }}>Add OTP</Button>
+    <Button size="sm" color="link" on:click={() => editingCredential = {
+        kind: 'Sso',
+        email: '',
+    }}>Add SSO</Button>
+</div>
+
+<div class="list-group list-group-flush mb-3">
+    {#each user.credentials as credential}
+    <div class="list-group-item credential">
+        {#if credential.kind === 'Password'}
+            <Fa fw icon={faKeyboard} />
+            <span class="type">Password</span>
+        {/if}
+        {#if credential.kind === 'PublicKey'}
+            <Fa fw icon={faKey} />
+            <span class="type">Public key</span>
+            <span class="text-muted ms-2">{abbreviatePublicKey(credential.key)}</span>
+        {/if}
+        {#if credential.kind === 'Totp'}
+            <Fa fw icon={faMobileScreen} />
+            <span class="type">One-time password</span>
+        {/if}
+        {#if credential.kind === 'Sso'}
+            <Fa fw icon={faIdBadge} />
+            <span class="type">Single sign-on</span>
+            <span class="text-muted ms-2">
+                {credential.email}
+                {#if credential.provider} ({credential.provider}){/if}
+            </span>
+        {/if}
+
         <span class="ms-auto"></span>
-        <Button size="sm" color="link" on:click={() => editingCredential = {
-            kind: 'Password',
-            hash: '',
-        }}>Add password</Button>
-        <Button size="sm" color="link" on:click={() => editingCredential = {
-            kind: 'PublicKey',
-            key: '',
-        }}>Add public key</Button>
-        <Button size="sm" color="link" on:click={() => editingCredential = {
-            kind: 'Totp',
-            key: [],
-        }}>Add OTP</Button>
-        <Button size="sm" color="link" on:click={() => editingCredential = {
-            kind: 'Sso',
-            email: '',
-        }}>Add SSO</Button>
+        <a
+            class="ms-2"
+            href={''}
+            onclick={e => {
+                editingCredential = credential
+                e.preventDefault()
+            }}>
+            Change
+        </a>
+        <a
+        class="ms-2"
+        href={''}
+        onclick={e => {
+            deleteCredential(credential)
+            e.preventDefault()
+        }}>
+        Delete
+    </a>
+</div>
+{/each}
+</div>
+
+<h4>Auth policy</h4>
+<div class="list-group list-group-flush mb-3">
+    {#each policyProtocols as protocol}
+    <div class="list-group-item">
+        <div>
+            <strong>{protocol.name}</strong>
+        </div>
+        {#if possibleCredentials[protocol.id]}
+        {@const _possibleCredentials = assertDefined(possibleCredentials[protocol.id])}
+        <AuthPolicyEditor
+        user={user}
+        bind:value={policy!}
+        possibleCredentials={_possibleCredentials}
+        protocolId={protocol.id}
+        />
+        {/if}
     </div>
+    {/each}
+</div>
 
-    <div class="list-group list-group-flush mb-3">
-        {#each user.credentials as credential}
-            <div class="list-group-item credential">
-                {#if credential.kind === 'Password'}
-                    <Fa fw icon={faKeyboard} />
-                    <span class="type">Password</span>
-                {/if}
-                {#if credential.kind === 'PublicKey'}
-                    <Fa fw icon={faKey} />
-                    <span class="type">Public key</span>
-                    <span class="text-muted ms-2">{abbreviatePublicKey(credential.key)}</span>
-                {/if}
-                {#if credential.kind === 'Totp'}
-                    <Fa fw icon={faMobileScreen} />
-                    <span class="type">One-time password</span>
-                {/if}
-                {#if credential.kind === 'Sso'}
-                    <Fa fw icon={faIdBadge} />
-                    <span class="type">Single sign-on</span>
-                    <span class="text-muted ms-2">
-                        {credential.email}
-                        {#if credential.provider} ({credential.provider}){/if}
-                    </span>
-                {/if}
-
-                <span class="ms-auto"></span>
-                <a
-                    class="ms-2"
-                    href={''}
-                    on:click|preventDefault={() =>
-                        editingCredential = credential
-                    }>
-                    Change
-                </a>
-                <a
-                    class="ms-2"
-                    href={''}
-                    on:click|preventDefault={() => deleteCredential(credential)}>
-                    Delete
-                </a>
-            </div>
-        {/each}
-    </div>
-
-    <h4>Auth policy</h4>
-    <div class="list-group list-group-flush mb-3">
-        {#each policyProtocols as protocol}
-            <div class="list-group-item">
-                <div>
-                    <strong>{protocol.name}</strong>
-                </div>
-                {#if possibleCredentials[protocol.id]}
-                    {@const _possibleCredentials = assertDefined(possibleCredentials[protocol.id])}
-                    <AuthPolicyEditor
-                        user={user}
-                        bind:value={policy}
-                        possibleCredentials={_possibleCredentials}
-                        protocolId={protocol.id}
-                    />
-                {/if}
-            </div>
-        {/each}
-    </div>
-
-    <h4 class="mt-4">User roles</h4>
-    <div class="list-group list-group-flush mb-3">
-        {#each allRoles as role}
-            <label
-                for="role-{role.id}"
-                class="list-group-item list-group-item-action d-flex align-items-center"
-            >
-                <Input
-                    id="role-{role.id}"
-                    class="mb-0 me-2"
-                    type="switch"
-                    on:change={() => toggleRole(role)}
-                    checked={roleIsAllowed[role.id]} />
-                <div>{role.name}</div>
-            </label>
-        {/each}
-    </div>
-
+<h4 class="mt-4">User roles</h4>
+<div class="list-group list-group-flush mb-3">
+    {#each allRoles as role}
+        <label
+            for="role-{role.id}"
+            class="list-group-item list-group-item-action d-flex align-items-center"
+        >
+            <Input
+            id="role-{role.id}"
+            class="mb-0 me-2"
+            type="switch"
+            on:change={() => toggleRole(role)}
+            checked={roleIsAllowed[role.id]} />
+            <div>{role.name}</div>
+        </label>
+    {/each}
+</div>
+{/if}
 {/await}
 
 {#if error}
@@ -266,7 +276,7 @@ function assertDefined<T>(value: T|undefined): T {
 {#if editingCredential}
 <UserCredentialModal
     credential={editingCredential}
-    username={user.username}
+    username={user!.username}
     save={saveCredential}
     cancel={() => editingCredential = undefined}
 />
