@@ -4,18 +4,19 @@ import AsyncButton from 'common/AsyncButton.svelte'
 import ConnectionInstructions from 'common/ConnectionInstructions.svelte'
 import { TargetKind } from 'gateway/lib/api'
 import { link } from 'svelte-spa-router'
-import { Alert, FormGroup } from '@sveltestrap/sveltestrap'
+import { FormGroup } from '@sveltestrap/sveltestrap'
 import { firstBy } from 'thenby'
 import { stringifyError } from 'common/errors'
+import Alert from 'common/Alert.svelte'
 
-let error: string|null = null
-let targets: Target[]|undefined
-let users: User[]|undefined
-let selectedTarget: Target|undefined
-let selectedUser: User|undefined
-let selectedExpiry: string|undefined
-let selectedNumberOfUses: number|undefined
-let result: TicketAndSecret|undefined
+let error: string|null = $state(null)
+let targets: Target[]|undefined = $state()
+let users: User[]|undefined = $state()
+let selectedTarget: Target|undefined = $state()
+let selectedUser: User|undefined = $state()
+let selectedExpiry: string|undefined = $state()
+let selectedNumberOfUses: number|undefined = $state()
+let result: TicketAndSecret|undefined = $state()
 
 async function load () {
     [targets, users] = await Promise.all([
