@@ -1,9 +1,9 @@
+import { mount } from 'svelte'
 import '../theme'
 import App from './App.svelte'
 
-new App({
+mount(App, {
     target: document.getElementById('app')!,
 })
 
-// eslint-disable-next-line @typescript-eslint/no-useless-empty-export
 export { }
