@@ -1,27 +1,31 @@
 <script lang="ts">
-import { faAngleLeft, faAngleRight } from '@fortawesome/free-solid-svg-icons'
-import Fa from 'svelte-fa'
-import { Pagination, PaginationItem, PaginationLink } from '@sveltestrap/sveltestrap'
+    import { faAngleLeft, faAngleRight } from '@fortawesome/free-solid-svg-icons'
+    import Fa from 'svelte-fa'
+    import { Pagination, PaginationItem, PaginationLink } from '@sveltestrap/sveltestrap'
 
-export let page = 0
-export let pageSize = 1
-export let total = 1
-
-let pages: (number|null)[] = []
-
-$: {
-    let i = 0
-    pages = []
-    let totalPages = Math.floor((total - 1) / pageSize + 1)
-    while (i < totalPages) {
-        if (i < 2 || i > totalPages - 3 || Math.abs(i - page) < 3) {
-            pages.push(i)
-        } else if (pages[pages.length - 1]) {
-            pages.push(null)
-        }
-        i++
+    interface Props {
+        page?: number;
+        pageSize?: number;
+        total?: number;
     }
-}
+
+    let { page = $bindable(0), pageSize = 1, total = 1 }: Props = $props()
+
+    let pages: (number|null)[] = $state([])
+
+    $effect(() => {
+        let i = 0
+        pages = []
+        let totalPages = Math.floor((total - 1) / pageSize + 1)
+        while (i < totalPages) {
+            if (i < 2 || i > totalPages - 3 || Math.abs(i - page) < 3) {
+                pages.push(i)
+            } else if (pages[pages.length - 1]) {
+                pages.push(null)
+            }
+            i++
+        }
+    })
 </script>
 
 <Pagination>
