@@ -1,5 +1,4 @@
 <script lang="ts">
-import { Alert } from '@sveltestrap/sveltestrap'
 import Router, { push, type RouteDetail } from 'svelte-spa-router'
 import { wrap } from 'svelte-spa-router/wrap'
 import { get } from 'svelte/store'
@@ -8,6 +7,7 @@ import ThemeSwitcher from 'common/ThemeSwitcher.svelte'
 import Logo from 'common/Logo.svelte'
 import DelayedSpinner from 'common/DelayedSpinner.svelte'
 import AuthBar from 'common/AuthBar.svelte'
+import Alert from 'common/Alert.svelte'
 
 let redirecting = false
 let serverInfoPromise = reloadServerInfo()
@@ -36,17 +36,17 @@ async function requireLogin (detail: RouteDetail) {
 
 const routes = {
     '/': wrap({
-        asyncComponent: () => import('./TargetList.svelte'),
+        asyncComponent: () => import('./TargetList.svelte') as any,
         props: {
             'on:navigation': () => redirecting = true,
         },
         conditions: [requireLogin],
     }),
     '/login': wrap({
-        asyncComponent: () => import('./Login.svelte'),
+        asyncComponent: () => import('./Login.svelte') as any,
     }),
     '/login/:stateId': wrap({
-        asyncComponent: () => import('./OutOfBandAuth.svelte'),
+        asyncComponent: () => import('./OutOfBandAuth.svelte') as any,
         conditions: [requireLogin],
     }),
 }
