@@ -4,15 +4,13 @@ import { faArrowRight } from '@fortawesome/free-solid-svg-icons'
 import ConnectionInstructions from 'common/ConnectionInstructions.svelte'
 import ItemList, { type LoadOptions, type PaginatedResponse } from 'common/ItemList.svelte'
 import { api, type TargetSnapshot, TargetKind } from 'gateway/lib/api'
-import { createEventDispatcher } from 'svelte'
 import Fa from 'svelte-fa'
-import { Modal, ModalBody, ModalHeader } from '@sveltestrap/sveltestrap'
+import { Modal, ModalBody } from '@sveltestrap/sveltestrap'
 import { serverInfo } from './lib/store'
 import { firstBy } from 'thenby'
+import ModalHeader from 'common/ModalHeader.svelte'
 
-const dispatch = createEventDispatcher()
-
-let selectedTarget: TargetSnapshot|undefined
+let selectedTarget: TargetSnapshot|undefined = $state()
 
 function loadTargets (options: LoadOptions): Observable<PaginatedResponse<TargetSnapshot>> {
     return from(api.getTargets({ search: options.search })).pipe(
@@ -41,52 +39,53 @@ function selectTarget (target: TargetSnapshot) {
 }
 
 function loadURL (url: string) {
-    dispatch('navigation')
     location.href = url
 }
 
 </script>
 
 <ItemList load={loadTargets} showSearch={true}>
-    <a
-        slot="item" let:item={target}
-        class="list-group-item list-group-item-action target-item"
-        href={
-            target.kind === TargetKind.WebAdmin
-                ? '/@warpgate/admin'
-                : target.kind === TargetKind.Http
-                    ? `/?warpgate-target=${target.name}`
-                    : '/@warpgate/admin'
-        }
-        on:click|preventDefault={e => {
-            if (e.metaKey || e.ctrlKey) {
-                return
+    {#snippet item({ item: target })}
+        <a
+            class="list-group-item list-group-item-action target-item"
+            href={
+                target.kind === TargetKind.WebAdmin
+                    ? '/@warpgate/admin'
+                    : target.kind === TargetKind.Http
+                        ? `/?warpgate-target=${target.name}`
+                        : '/@warpgate/admin'
             }
-            selectTarget(target)
-        }}
-    >
-        <span class="me-auto">
-            {#if target.kind === TargetKind.WebAdmin}
-                Manage Warpgate
-            {:else}
-                {target.name}
+            onclick={e => {
+                if (e.metaKey || e.ctrlKey) {
+                    return
+                }
+                e.preventDefault()
+                selectTarget(target)
+            }}
+        >
+            <span class="me-auto">
+                {#if target.kind === TargetKind.WebAdmin}
+                    Manage Warpgate
+                {:else}
+                    {target.name}
+                {/if}
+            </span>
+            <small class="protocol text-muted ms-auto">
+                {#if target.kind === TargetKind.Ssh}
+                    SSH
+                {/if}
+                {#if target.kind === TargetKind.MySql}
+                    MySQL
+                {/if}
+                {#if target.kind === TargetKind.Postgres}
+                    PostgreSQL
+                {/if}
+            </small>
+            {#if target.kind === TargetKind.Http || target.kind === TargetKind.WebAdmin}
+                <Fa icon={faArrowRight} fw />
             {/if}
-        </span>
-        <small class="protocol text-muted ms-auto">
-            {#if target.kind === TargetKind.Ssh}
-                SSH
-            {/if}
-            {#if target.kind === TargetKind.MySql}
-                MySQL
-            {/if}
-            {#if target.kind === TargetKind.Postgres}
-                PostgreSQL
-            {/if}
-        </small>
-        {#if target.kind === TargetKind.Http || target.kind === TargetKind.WebAdmin}
-            <Fa icon={faArrowRight} fw />
-        {/if}
-    </a>
+        </a>
+    {/snippet}
 </ItemList>
 
 <Modal isOpen={!!selectedTarget} toggle={() => selectedTarget = undefined}>
